Model property naming style generic for scala
